+++ lib/cuke_modeler/adapters/gherkin_16_adapter.rb
@@ -1,0 +1,13 @@
+require_relative 'gherkin_9_adapter'
+
+
+module CukeModeler
+
+  # NOT A PART OF THE PUBLIC API
+  # An adapter that can convert the output of version 16.x of the *cucumber-gherkin* gem into input that is consumable
+  # by this gem.
+
+  class Gherkin16Adapter < Gherkin9Adapter
+
+  end
+end